validate authorization on post article

+++ api/article/article.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"strconv"
 	"strings"
 
 	"github.com/akinaru-lu/elesion"
@@ -12,7 +11,6 @@ import (
 
 	"lmm/api/db"
 	"lmm/api/user"
-	"lmm/api/utils/token"
 )
 
 type Article struct {
@@ -70,26 +68,11 @@ func getArticles(values db.Values) ([]Article, error) {
 }
 
 // NewArticle post new article to the user given by url path
-// POST /users/:user/articles
+// POST /articles
 func NewArticle(c *elesion.Context) {
-	// userIDStr := c.Params.ByName("user")
-	// userID, err := strconv.ParseInt(userIDStr, 10, 64)
-	// if err != nil {
-	// 	c.Status(http.StatusBadRequest).String("invalid user id: " + userIDStr)
-	// 	return
-	// }
-	accessToken := c.Request.Header.Get("Authorization")
-	originToken, err := token.Decode(accessToken)
+	usr, err := user.CheckAuth(c.Request.Header.Get("Authorization"))
 	if err != nil {
-		c.Status(http.StatusUnauthorized).String("Unauthorized, invalid token").Error(err.Error())
-		return
-	}
-
-	values := db.NewValues()
-	values["token"] = originToken
-	usr, err := user.GetUser(values)
-	if err != nil {
-		c.Status(http.StatusUnauthorized).String("Unauthorized").Error(err.Error())
+		c.Status(http.StatusUnauthorized).String(err.Error())
 		return
 	}
 
@@ -156,25 +139,16 @@ func NewTestArticle() (*Article, *user.UserProfile) {
 */
 
 // UpdateArticle update the article where user name and article id are matched
-// PUT /users/:user/articles/:article
-// TODO params 'user' stands for user id
+// PUT /articles
 func UpdateArticle(c *elesion.Context) {
-	userIDStr := c.Params.ByName("user")
-	userID, err := strconv.ParseInt(userIDStr, 10, 64)
+	usr, err := user.CheckAuth(c.Request.Header.Get("Authorization"))
 	if err != nil {
-		c.Status(http.StatusBadRequest).String("invalid user id: " + userIDStr)
-		return
-	}
-
-	articleIDStr := c.Params.ByName("article")
-	articleID, err := strconv.ParseInt(articleIDStr, 10, 64)
-	if err != nil || articleID <= 0 {
-		c.Status(http.StatusBadRequest).String("invalid article id: " + articleIDStr)
+		c.Status(http.StatusUnauthorized).String(err.Error())
 		return
 	}
 
 	body := Article{}
-	body.User = userID
+	body.User = usr.ID
 	err = json.NewDecoder(c.Request.Body).Decode(&body)
 	if err != nil {
 		c.Status(http.StatusBadRequest).Error(err.Error()).String("invalid body")
@@ -182,57 +156,97 @@ func UpdateArticle(c *elesion.Context) {
 	}
 	defer c.Request.Body.Close()
 
-	err = updateArticle(articleID, body)
-	if err != nil {
+	err = updateArticle(body)
+
+	switch err {
+	case nil:
+		c.Status(http.StatusOK).String("success")
+	case db.ErrNoChange:
+		c.Status(http.StatusAccepted).String("no change")
+	case db.ErrNoRows:
+		c.Status(http.StatusNotFound).String(fmt.Sprintf("no such article: %d", body.ID))
+	default:
 		c.Status(http.StatusInternalServerError).Error(err.Error()).String(err.Error())
-		return
 	}
-	c.Status(http.StatusOK).String("success")
 }
 
-func updateArticle(id int64, article Article) error {
+func updateArticle(article Article) error {
 	d := db.UseDefault()
 	defer d.Close()
 
+	ok, err := d.Exists("SELECT 1 FROM article WHERE id = ? AND user = ?", article.ID, article.User)
+	if err != nil {
+		return err
+	}
+	if !ok {
+		return db.ErrNoRows
+	}
+
 	res, err := d.Exec("UPDATE article SET title = ?, text = ? WHERE id = ? AND user = ?",
-		article.Title, article.Text, id, article.User,
+		article.Title, article.Text, article.ID, article.User,
 	)
+	if err != nil {
+		return err
+	}
 	if rows, err := res.RowsAffected(); err != nil {
 		return err
-	} else if rows != 1 {
-		return errors.WithCaller("rows affected should be 1", 2)
+	} else if rows == 0 {
+		return db.ErrNoChange
+	} else if rows > 1 {
+		return errors.Newf("rows affected should be larger than 1 but got ", rows)
 	}
-	return err
+	return nil
 }
 
 // DeleteArticle delete the article where user name and article id are matched
-// DELETE /users/:user/articles/:article
+// DELETE /articles
 func DeleteArticle(c *elesion.Context) {
-	user := c.Params.ByName("user")
-	articleStr := c.Params.ByName("article")
-	article, err := strconv.ParseInt(articleStr, 10, 64)
+	usr, err := user.CheckAuth(c.Request.Header.Get("Authorization"))
 	if err != nil {
-		c.Status(http.StatusBadRequest).String("invalid article id: " + articleStr)
+		c.Status(http.StatusUnauthorized).String(err.Error())
 		return
 	}
 
-	err = deleteArticle(user, article)
+	article := Article{}
+	err = json.NewDecoder(c.Request.Body).Decode(&article)
 	if err != nil {
-		c.Status(http.StatusInternalServerError).Error(err.Error()).String(err.Error())
+		c.Status(http.StatusBadRequest).String("invald body")
 		return
 	}
-	c.Status(http.StatusOK).String("success")
+
+	err = deleteArticle(usr.ID, article.ID)
+	switch err {
+	case nil:
+		c.Status(http.StatusOK).String("success")
+	case db.ErrNoRows:
+		c.Status(http.StatusNotFound).Stringf("article not found")
+	default:
+		c.Status(http.StatusInternalServerError).Error(err.Error()).String(err.Error())
+	}
 }
 
-func deleteArticle(user string, id int64) error {
+func deleteArticle(user, id int64) error {
 	d := db.UseDefault()
 	defer d.Close()
 
+	ok, err := d.Exists("SELECT 1 FROM article WHERE id = ? AND user = ?", id, user)
+	if err != nil {
+		return err
+	}
+	if !ok {
+		return db.ErrNoRows
+	}
+
 	result, err := d.Exec("DELETE FROM article WHERE id = ? AND user = ?", id, user)
+	if err != nil {
+		return err
+	}
 	if rows, err := result.RowsAffected(); err != nil {
 		return err
-	} else if rows != 1 {
+	} else if rows == 0 {
+		return db.ErrNoChange
+	} else if rows > 1 {
 		return errors.Newf("rows affected shoulb be 1 but got %d", rows)
 	}
-	return err
+	return nil
 }
